Simplifying well-known conf path generation
