+++ sso/provider-generic/src/main/resources/sso-generic-setting.properties
@@ -17,4 +17,3 @@
 #sso.generic.openid.server.endpoint.auth=http://localhost:9090/auth/realms/kapua/protocol/openid-connect/auth
 #sso.generic.openid.server.endpoint.logout=http://localhost:9090/auth/realms/kapua/protocol/openid-connect/logout
 #sso.generic.openid.server.endpoint.token=http://localhost:9090/auth/realms/kapua/protocol/openid-connect/token
-#sso.generic.openid.conf.wellknown.path=/.well-known/openid-configuration
