+++ sso/provider/src/main/resources/sso-setting.properties
@@ -11,8 +11,7 @@
 #
 ###############################################################################
 
-sso.openid.conf.path=.well-known/openid-configuration
-
 # properties with example values:
 #sso.provider=keycloak
 #sso.openid.client.id=console
+#sso.openid.conf.path=.well-known/openid-configuration
